// src/api/schema/transactionStatusEnum.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI spec version: 0.0.0
 */

/**
 * * `pending` - pending
* `approved` - approved
* `rejected` - rejected
* `active` - active
* `returning` - returning
* `completed` - completed
 */
export type TransactionStatusEnum = typeof TransactionStatusEnum[keyof typeof TransactionStatusEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TransactionStatusEnum = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  active: 'active',
  returning: 'returning',
  completed: 'completed',
} as const;